pyproject.toml
[project]
name = "looker-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.30.0",
    "lkml>=1.3.7",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "ruamel-yaml>=0.18.10",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
looker-loader = "src.cli:main"
